+++ src/assets/ba_data/python/baenv.py
@@ -56,8 +56,8 @@ logger = logging.getLogger('ba.env')
 
 # Build number and version of the ballistica binary we expect to be
 # using.
-TARGET_BALLISTICA_BUILD = 22597
-TARGET_BALLISTICA_VERSION = '1.7.53'
+TARGET_BALLISTICA_BUILD = 22599
+TARGET_BALLISTICA_VERSION = '1.7.54'
 
 
 @dataclass