+++ src/ballistica/shared/ballistica.cc
@@ -44,8 +44,8 @@ auto main(int argc, char** argv) -> int {
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kEngineBuildNumber = 22597;
-const char* kEngineVersion = "1.7.53";
+const int kEngineBuildNumber = 22599;
+const char* kEngineVersion = "1.7.54";
 const int kEngineApiVersion = 9;
 
 #if BA_MONOLITHIC_BUILD